Enforce Checkstyle ruleset

+++ kafka-webview-ui/src/main/java/com/darksci/kafka/webview/ui/manager/plugin/UploadManager.java
@@ -96,6 +96,11 @@ public class UploadManager {
             return true;
         }
 
+        // Only remove files
+        if (!fullOutputPath.toFile().isFile()) {
+            return false;
+        }
+
         try {
             Files.delete(fullOutputPath);
         } catch (IOException ex) {
